Responsiveness adjusted for CategoryDetail component
