+++ src/components/App/App.css
@@ -169,17 +169,6 @@ input.cocktail-box {
   outline: none;
 }
 
-.category-details{
-  padding: 4rem;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.category-details-info {
-  color: #645b92;
-}
-
 .drinks-list {
   display: flex;
   justify-content: space-between;
@@ -203,37 +192,6 @@ input.cocktail-box {
   font-weight: bolder;
 }
 
-.category-card{
-  display: flex;
-  flex-direction: column;
-  background-color: white;
-  max-height: 100%;
-  width: 270px;
-  box-shadow: 0px 1px 5px #555;
-  border-radius: 20px;
-  overflow: hidden;
-  opacity: 0.8;
-  justify-content: center;
-}
-
-.category-card h3  {
-  margin-bottom: 20px;
-  margin-top: 80px;
-}
-
-.category-card p {
-  margin-top: 10px;
-  margin-bottom: 5px;
-  margin-left: 20px;
-  margin-right: 20px;
-  overflow-y: scroll;
-  height: 120px;
-}
-
-.category-card.view-more-btn {
-  padding-bottom: 10px;
-  margin-top: -5px;
-}
 
 .image-category {
   height: 250px;
