+++ src/features/categories/CategoryCard/CategoryCard.module.css
@@ -190,7 +190,7 @@
     }
 }
 
-@media (orientation: landscape) and (min-width: 1100px) {
+@media (orientation: landscape) and (min-width: 1100px) and (max-width: 1300px) {
     .card {
         padding: 1rem 1rem 0;
     }
