+++ src/features/cocktails/CocktailsContainer/CocktailsContainer.module.css
@@ -60,7 +60,7 @@
     width: 100%;
     display: grid;
     padding: 0 1rem;
-    grid-template-columns: repeat(5, 1fr);
+    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
     grid-gap: 3rem;
     text-align: center;
 }
@@ -89,10 +89,12 @@
 @media (min-width: 340px) and (max-width: 430px) {
     .container {
         gap: 1rem;
-        padding: 0 3rem;
+        margin-top: 3rem;
+        padding: 0 2rem;
     }
     .cocktailsGrid {
-        grid-template-columns: repeat(1, 1fr);
+        padding: 0;
+        grid-gap: 1rem;
     }
     .addCocktail {
         width: 50%;
@@ -109,11 +111,13 @@
 @media (min-width: 430px) and (max-width: 600px) {
     .container {
         gap: 1rem;
+        margin-top: 5rem;
         padding: 0 1rem;
     }
     .cocktailsGrid {
-        grid-template-columns: repeat(2, 1fr);
+        grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
         grid-gap: 1rem;
+        padding: 0;
     }
     .addCocktail {
         width: 30%;
@@ -133,11 +137,13 @@
 @media (min-width: 600px) and (max-width: 840px) {
     .container {
         gap: 0.5rem;
+        margin-top: 5rem;
         padding: 0 1rem;
     }
     .cocktailsGrid {
-        grid-template-columns: repeat(3, 1fr);
+        grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
         grid-gap: 1rem;
+        padding: 0;
     }
     .addCocktail {
         width: 20%;
@@ -157,11 +163,13 @@
 @media (min-width: 840px) and (max-width: 1180px) {
     .container {
         gap: 1rem;
+        margin-top: 5rem;
         padding: 0 1rem;
     }
     .cocktailsGrid {
-        grid-template-columns: repeat(4, 1fr);
+        grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
         grid-gap: 1rem;
+        padding: 0;
     }
     .addCocktail {
         width: 15%;
@@ -198,20 +206,37 @@
 @media (orientation: landscape) and (min-width: 600px) and (max-width: 800px) {
     .container {
         gap: 1rem;
+        margin-top: 3rem;
         padding: 0 1.5rem;
     }
     .cocktailsGrid {
-        grid-template-columns: repeat(3, 1fr);
+        grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+        padding: 0;
     }
 }
 
-@media (orientation: landscape) and (min-width: 800px) and (max-width: 1100px) {
+@media (orientation: landscape) and (min-width: 800px) and (max-width: 950px) {
     .container {
         gap: 0.5rem;
+        margin-top: 2rem;
         padding: 0 1rem;
     }
     .cocktailsGrid {
-        grid-template-columns: repeat(4, 1fr);
+        grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+        padding: 0;
+    }
+}
+
+@media (orientation: landscape) and (min-width: 950px) and (max-width: 1100px) {
+    .container {
+        gap: 0.5rem;
+        margin-top: 4rem;
+        padding: 0 1rem;
+    }
+
+    .cocktailsGrid {
+        grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+        padding: 0;
     }
 }
 
@@ -221,6 +246,6 @@
         padding: 0 1rem;
     }
     .cocktailsGrid {
-        grid-template-columns: repeat(5, 1fr);
+        padding: 0 1rem;
     }
 }
